+++ src/sql_bridge_emysql.erl
@@ -1,5 +1,7 @@
 -module(sql_bridge_emysql).
 -behaviour(sql_bridge_adapter).
+-include_lib("emysql/include/emysql.hrl").
+-define(IS_DATE(Type), Type==?FIELD_TYPE_DATETIME orelse Type==?FIELD_TYPE_TIMESTAMP).
 
 -export([
 	start/0,
@@ -63,19 +65,24 @@ maybe_replace_tokens(Q, ParamList) ->
 %% @doc Format the results from emysql as a list of Types
 format_result(Type,Res) ->
     Json = emysql:as_json(Res),
+    Fields = get_field_types(Res),
+    DateFields = [?IS_DATE(F) || F <- Fields],
 	Stringify = sql_bridge_utils:stringify_binaries(),
     case Type of
         list ->
-            format_list_result(Stringify, Json);
+            format_list_result(DateFields, Stringify, Json);
         tuple ->
-            format_tuple_result(Stringify, Json);
+            format_tuple_result(DateFields, Stringify, Json);
         proplist ->
-            format_proplist_result(Stringify, Json);
+            format_proplist_result(DateFields, Stringify, Json);
         dict ->
-            format_dict_result(Stringify, Json);
+            format_dict_result(DateFields, Stringify, Json);
 		map ->
-			format_map_results(Stringify, Json)
+			format_map_results(DateFields, Stringify, Json)
     end.
+
+get_field_types(#result_packet{field_list=Fs}) ->
+    [F#field.type || F <- Fs].
 
 -spec format_key(K :: sql_bridge:field()) -> atom().
 %% @doc Normalize field values into atoms
@@ -86,39 +93,51 @@ format_key(K) when is_binary(K) ->
 format_key(K) when is_list(K) ->
     list_to_atom(K).
 
--spec format_list_result(Stringify :: boolean(), Json :: sql_bridge:json()) -> [list()].
-format_list_result(Stringify, Json) ->
+-spec format_list_result(DateFields :: [boolean()], Stringify :: boolean(), Json :: sql_bridge:json()) -> [list()].
+format_list_result(DateFields, Stringify, Json) ->
     [
-        [format_value(Stringify, Value) || {_,Value} <- Row]
+        [format_value(IsDate, Stringify, Value) || {IsDate, {_,Value}} <- lists:zip(DateFields, Row)]
     || Row <- Json].
 
--spec format_tuple_result(Stringify :: boolean(), Json :: json()) -> [tuple()].
-format_tuple_result(Stringify, Json) ->
-    [list_to_tuple(Row) || Row <- format_list_result(Stringify, Json)].
+-spec format_tuple_result(DateFields :: [boolean()], Stringify :: boolean(), Json :: json()) -> [tuple()].
+format_tuple_result(DateFields, Stringify, Json) ->
+    [list_to_tuple(Row) || Row <- format_list_result(DateFields, Stringify, Json)].
 
--spec format_proplist_result(Stringify :: boolean(), Json :: json()) -> [sql_bridge:proplist()].
-format_proplist_result(Stringify, Json) ->
+-spec format_proplist_result(DateFields :: [boolean()], Stringify :: boolean(), Json :: json()) -> [sql_bridge:proplist()].
+format_proplist_result(DateFields, Stringify, Json) ->
     [
-        [{format_key(F), format_value(Stringify, V)} || {F,V} <- Row]
+        [{format_key(F), format_value(IsDate, Stringify, V)} || {IsDate, {F,V}} <- lists:zip(DateFields, Row)]
     || Row <-Json].
 
--spec format_dict_result(Stringify :: boolean(), Json :: json()) -> [sql_bridge:t_dict()].
-format_dict_result(Stringify, Json) ->
-    [dict:from_list(PL) || PL <- format_proplist_result(Stringify, Json)].
+-spec format_dict_result(DateFields :: [boolean()], Stringify :: boolean(), Json :: json()) -> [sql_bridge:t_dict()].
+format_dict_result(DateFields, Stringify, Json) ->
+    [dict:from_list(PL) || PL <- format_proplist_result(DateFields, Stringify, Json)].
 
--spec format_map_results(Stringify :: boolean(), Json :: json()) -> [map()].
-format_map_results(Stringify, Json) ->
-    [maps:from_list(PL) || PL <- format_proplist_result(Stringify, Json)].
+-spec format_map_results(DateFields :: [boolean()], Stringify :: boolean(), Json :: json()) -> [map()].
+format_map_results(DateFields, Stringify, Json) ->
+    [maps:from_list(PL) || PL <- format_proplist_result(DateFields, Stringify, Json)].
 
 schema_db_column() ->
 	"table_schema".
 
-format_value(_, null) ->
+format_value(_, _, null) ->
 	undefined;
-format_value(true, V) when is_binary(V) ->
+format_value(true, Stringify, V) ->
+    Return = case Stringify of
+        true -> list;
+        false -> binary
+    end,
+    {match, [D,T]} = re:run(V, "^(\\d{4}-\\d{2}-\\d{2})T(\\d{2}:\\d{2}:\\d{2})Z$",[{capture, all_but_first, Return}]),
+    case Stringify of
+        true -> D ++ " " ++ T;
+        false -> <<D/binary," ",T/binary>>
+    end;
+format_value(_, true, V) when is_binary(V) ->
 	sql_bridge_utils:binary_to_string(V);
-format_value(_, V) ->
-	V.
+format_value(_, _, V) when is_tuple(V) ->
+    sql_bridge_utils:format_datetime(V);
+format_value(_, _, V) ->
+    V.
 
 -spec encode(V :: any()) -> binary().
 %% @doc Safely encodes text for insertion into a query.  Replaces the atoms