+++ src/sql_bridge_mysql_otp.erl
@@ -62,7 +62,14 @@ query(Type, DB, Q, ParamList) ->
 	try query_catched(Type, DB, Q, ParamList)
 	catch
 		exit:{noproc, _} ->
-			{error, no_pool}
+            {error, no_pool};
+        exit:{{{badmatch,{error,closed}}, _}, _} ->
+            {error, disconnected};
+        exit:{{{case_clause,{error,closed}}, _}, _} ->
+            {error, disconnected};
+        E:T ->
+            error_logger:info_msg("Unhandled Error: ~p:~p", [E,T]),
+            throw(unhandled_error)
 	end.
 
 query_catched(Type, DB, Q, ParamList) ->
@@ -83,7 +90,15 @@ query_catched(Type, DB, Q, ParamList) ->
 mysql_query(Worker, Q, []) ->
 	mysql:query(Worker, Q);
 mysql_query(Worker, Q, ParamList) ->
-	mysql:query(Worker, Q, ParamList).
+    ParamList2 = pre_encode(ParamList),
+	mysql:query(Worker, Q, ParamList2).
+
+pre_encode(List) ->
+    lists:map(fun(true) -> 1;
+                 (false) -> 0;
+                 (undefined) -> null;
+                 (X) -> X
+              end, List).
 
 maybe_replace_tokens(Q, ParamList) ->
 	case sql_bridge_utils:replacement_token() of
@@ -127,7 +142,12 @@ format_lists(Rows) ->
 	end.
 
 format_list(Row) when is_list(Row) ->
-	[sql_bridge_stringify:maybe_string(V) || V <- Row].
+	[normalize_value(V) || V <- Row].
+
+normalize_value(V) when is_tuple(V) ->
+    sql_bridge_utils:format_datetime(V);
+normalize_value(V) ->
+	sql_bridge_stringify:maybe_string(V).
 
 format_proplists(Columns, Rows) ->
 	ColNames = extract_colnames(Columns),
@@ -143,7 +163,7 @@ make_dict(Cols, Row) when is_list(Row) ->
 make_dict([], [], Dict) ->
 	Dict;
 make_dict([Col|Cols], [Val|Vals], Dict) ->
-	Val2 = sql_bridge_stringify:maybe_string(Val),
+	Val2 = normalize_value(Val),
 	NewDict = dict:store(Col, Val2, Dict),
 	make_dict(Cols, Vals, NewDict).
 	
@@ -153,7 +173,7 @@ extract_colnames(Columns) ->
 make_proplist(Columns, Row) when is_tuple(Row) ->
 	make_proplist(Columns, tuple_to_list(Row));
 make_proplist([Col|Cols], [Val|Vals]) ->
-	Val2 = sql_bridge_stringify:maybe_string(Val),
+	Val2 = normalize_value(Val),
 	[{Col, Val2} | make_proplist(Cols, Vals)];
 make_proplist([], []) ->
 	[].
@@ -169,7 +189,7 @@ make_map(Cols, Row) ->
 make_map([], [], Map) ->
 	Map;
 make_map([Col|Cols],[Val|Vals], Map) ->
-	Val2 = sql_bridge_stringify:maybe_string(Val),
+	Val2 = normalize_value(Val),
 	NewMap = maps:put(Col, Val2, Map),
 	make_map(Cols, Vals, NewMap).