+++ src/sql_bridge_utils.erl
@@ -2,6 +2,8 @@
 -module(sql_bridge_utils).
 -export([
     get_env/2,
+    record_handler/0,
+    convert_record/1,
     start_poolboy_pool/3,
     to_string/1,
     with_poolboy_pool/2,
@@ -16,7 +18,9 @@
     create_placeholder/1,
     to_atom/1,
     checkout_pool/1,
-    checkin_pool/1
+    checkin_pool/1,
+    record_to_proplist/2,
+    format_datetime/1
 ]).
 
 replacement_token() ->
@@ -29,6 +33,17 @@ replacement_token() ->
 
 stringify_binaries() ->
     get_env(stringify_binaries, false).
+
+record_handler() ->
+    case get_env(record_handler, undefined) of
+        undefined -> throw("No record handler defined");
+        F when is_function(F) -> F;
+        {Mod, Fun} -> fun Mod:Fun/1
+    end.
+
+convert_record(Record) ->
+    Handler = record_handler(),
+    Handler(Record).
 
 start_poolboy_pool(Name, WorkerArgs, WorkerModule) ->
     Size = get_env(connections_per_pool, 10),
@@ -56,7 +71,6 @@ with_poolboy_pool(DB, Fun) ->
             poolboy:checkin(DB, Worker),
             Return;
         {ok, Worker} ->
-            io:format("~p in trans: ~p~n",[self(), Worker]),
             _Return = Fun(Worker)
     end.
 
@@ -202,3 +216,36 @@ to_atom(A) when is_list(A) ->
     list_to_atom(A);
 to_atom(A) when is_atom(A) ->
     A.
+
+record_to_proplist(Record, FieldMap) ->
+    NumberedFields = lists:zip(lists:seq(2, length(FieldMap)+1), FieldMap),
+    lists:map(fun({ElNum, Field}) ->
+        Value = element(ElNum, Record),
+        {Field, Value}
+    end, NumberedFields).
+
+format_datetime({{Y,M,D},{H,I,S}}) when is_float(S) ->
+    format_datetime({{Y,M,D},{H,I,erlang:trunc(S)}});
+format_datetime({{Y,M,D},{H,I,S}}) ->
+    to_bin_or_str(io_lib:format("~4..0B-~2..0B-~2..0B ~2..0B:~2..0B:~2..0B", [Y,M,D,H,I,S]));
+format_datetime({time, {H,I,S}}) ->
+    format_datetime({0, {H,I,S}});
+format_datetime({0, {H,I,S}}) when is_integer(S) ->
+    to_bin_or_str(io_lib:format("~2..0B:~2..0B:~2..0B", [H,I,S]));
+format_datetime({0, {H,I,S}}) when is_float(S) ->
+    format_datetime({0, {H,I,erlang:trunc(S)}});
+format_datetime({H,I,S}) when is_float(S) ->
+    format_datetime({0, {H,I,S}});
+format_datetime({Y,M,D}) ->
+	to_bin_or_str(io_lib:format("~4..0B-~2..0B-~2..0B", [Y,M,D])).
+
+to_bin_or_str(B) when is_binary(B) ->
+    case stringify_binaries() of
+        true -> sql_bridge_stringify:maybe_string(B);
+        false -> B
+    end;
+to_bin_or_str(L) when is_list(L) ->
+    case stringify_binaries() of
+        true -> lists:flatten(L);
+        false -> iolist_to_binary(L)
+    end.