+++ CHANGELOG.markdown
@@ -1,6 +1,17 @@
+## v0.3.0
+  * Normalize return values for date, time, timestamp, and decimal types.
+  * Fix `db:save` to properly decompose tables of the `database.tablename`
+    format.
+  * Better normalization of `true`, `false` and NULLs to `undefined`
+  * Add more tests
+
+## v0.2.0
+  * Add record handling
+  * Improve stability in mysql-otp driver
+
 ## v0.1.0
   * Add support for mysql-otp and epgsql drivers
   * Add battery of tests
-  * 
+
 ## Prior Versions
   * Supported Emysql driver only.